add motivation to Create order screen
